refactor(ROB-560): rename live n8n_* data services to role-based names
n8n_filled_orders_service -> filled_orders_service (execution_ledger reconciler + /invest/stocks)
n8n_pending_orders_service -> pending_orders_service (intraday_order_review)
n8n_market_context_service -> market_context_service
n8n_formatting -> order_brief_formatting
Pure rename, behavior unchanged. Dead n8n HTTP surface removal is 2b.
