+++ tests/services/research_run_safety_helpers.py
@@ -27,7 +27,7 @@ RESEARCH_RUN_FORBIDDEN_PREFIXES = [
     "app.services.kis_holdings_service",
     "app.services.upbit_websocket",
     "app.services.redis_token_manager",
-    "app.services.n8n_pending_orders_service",
+    "app.services.pending_orders_service",
     "app.services.n8n_pending_review_service",
     "app.mcp_server.tooling.order_execution",
     "app.mcp_server.tooling.orders_history",
@@ -47,7 +47,7 @@ NEWS_BRIEF_FORBIDDEN_PREFIXES = [
         "app.services.kis_websocket_internal",
         "app.services.kis_trading_contracts",
         "app.services.kis_holdings_service",
-        "app.services.n8n_pending_orders_service",
+        "app.services.pending_orders_service",
         "app.services.n8n_pending_review_service",
         "app.mcp_server.tooling.order_execution",
         "app.mcp_server.tooling.orders_history",
